+++ relay-server/src/endpoints/common.rs
@@ -11,7 +11,7 @@ use serde::Deserialize;
 use crate::envelope::{AttachmentType, Envelope, EnvelopeError, Item, ItemType, Items};
 use crate::service::ServiceState;
 use crate::services::buffer::{EnvelopeBuffer, ProjectKeyPair};
-use crate::services::outcome::{DiscardReason, Outcome};
+use crate::services::outcome::{DiscardItemType, DiscardReason, Outcome};
 use crate::services::processor::{BucketSource, MetricData, ProcessMetrics, ProcessingGroup};
 use crate::statsd::{RelayCounters, RelayHistograms};
 use crate::utils::{self, ApiErrorResponse, FormDataIter, ManagedEnvelope};
@@ -83,7 +83,7 @@ pub enum BadStoreRequest {
     #[error(
         "envelope exceeded size limits for type '{0}' (https://develop.sentry.dev/sdk/envelopes/#size-limits)"
     )]
-    Overflow(ItemType),
+    Overflow(DiscardItemType),
 
     #[error(
         "Sentry dropped data due to a quota or internal rate limit being reached. This will not affect your application. See https://docs.sentry.io/product/accounts/quotas/ for more information."
@@ -385,9 +385,7 @@ pub async fn handle_envelope(
     if let Err(offender) =
         utils::check_envelope_size_limits(state.config(), managed_envelope.envelope())
     {
-        managed_envelope.reject(Outcome::Invalid(DiscardReason::TooLarge(
-            (&offender).into(),
-        )));
+        managed_envelope.reject(Outcome::Invalid(DiscardReason::TooLarge(offender)));
         return Err(BadStoreRequest::Overflow(offender));
     }